// apps/creator-hub/modules/localization/gameStringTranslation/enums/EntryFilterOptions.ts
enum EntryFilterOptions {
  Translated = 'Translated',
  Untranslated = 'Untranslated',
  AutomaticTranslated = 'AutomaticTranslated',
  UserTranslated = 'UserTranslated',
  RecentlyAddedEntries = 'RecentlyAddedEntries',
  RecentlyModifiedTranslations = 'RecentlyModifiedTranslations',
  FeedbackAvailable = 'FeedbackAvailable',
  NoFeedback = 'NoFeedback',
}

export default EntryFilterOptions;